agents/Codex/requirements.txt
# Codex agent dependencies
pyyaml>=6.0.0
